Modify Files

+++ factorial.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int factorial(int x)
+double factorial(double x)
 {
 
     if (x == 1 || x == 0)
@@ -16,6 +16,6 @@ int main()
     int num;
     printf("Enter a number you want factorial of:\n");
     scanf("%d", &num);
-    printf("The factorial of %d is %d\n", num, factorial(num));
+    printf("The factorial of %d is %lf\n", num, factorial(num));
     return 0;
 }
